requirements.txt
streamlit
pandas
scikit-learn
matplotlib
seaborn
rdkit-pypi
numpy